Support to run local game.json
