+++ packages/headless-driver-runner-v1/package.json
@@ -22,13 +22,11 @@
   "typings": "lib/index.d.ts",
   "dependencies": {
     "@akashic/engine-files": "1.1.9",
-    "@akashic/headless-driver-runner": "0.4.0",
-    "node-fetch": "2.3.0"
+    "@akashic/headless-driver-runner": "0.4.0"
   },
   "devDependencies": {
-    "@types/node-fetch": "2.1.2",
     "prettier": "^1.17.0",
-    "rimraf": "2.6.3",
+    "rimraf": "^2.6.3",
     "tslint": "^5.16.0",
     "typescript": "^3.4.5"
   }
