+++ packages/headless-driver-runner-v2/src/platform/assets/NodeTextAsset.ts
@@ -1,10 +1,9 @@
 import { akashicEngine as g } from "@akashic/engine-files";
-import fetch from "node-fetch";
+import { loadFile } from "@akashic/headless-driver-runner";
 
 export class NodeTextAsset extends g.TextAsset {
 	_load(loader: g.AssetLoadHandler): void {
-		fetch(this.path, { method: "GET" })
-			.then(res => res.text())
+		loadFile<string>(this.path, {json: false})
 			.then(text => {
 				this.data = text;
 				return loader._onAssetLoad(this);
