+++ packages/headless-driver-runner/package.json
@@ -23,12 +23,12 @@
   "dependencies": {
     "@akashic/amflow": "0.2.2",
     "@akashic/trigger": "0.1.5",
-    "node-fetch": "2.3.0"
+    "node-fetch": "2.5.0"
   },
   "devDependencies": {
-    "@types/node-fetch": "2.1.2",
+    "@types/node-fetch": "2.3.3",
     "prettier": "^1.16.0",
-    "rimraf": "2.6.3",
+    "rimraf": "^2.6.3",
     "tslint": "^5.16.0",
     "typescript": "^3.4.5"
   }
