+++ packages/headless-driver-runner/src/index.ts
@@ -1,3 +1,4 @@
 export * from "./Runner";
 export * from "./Platform";
 export * from "./Looper";
+export * from "./utils";
