+++ packages/headless-driver-runner/src/utils.ts
@@ -1,0 +1,37 @@
+import { readFileSync } from "fs";
+import fetch from "node-fetch";
+import * as path from "path";
+
+export interface ReadFileOption {
+	encoding?: string;
+	json?: boolean;
+}
+
+/**
+ * テキストファイルの読み込みを行う。
+ *
+ * @param url url または path
+ */
+export async function loadFile(url: string): Promise<string>;
+
+/**
+ * ファイルの読み込みを行う。
+ *
+ * @param url url または path
+ * @param opt オプション
+ */
+export async function loadFile<T>(url: string, opt?: ReadFileOption): Promise<T>;
+
+export async function loadFile<T>(url: string, opt?: ReadFileOption): Promise<T> {
+	if (isHttpProtocol(url)) {
+		const res = await fetch(url, { method: "GET" });
+		return opt.json ? res.json() : res.text();
+	} else {
+		const str = readFileSync(path.resolve(this.assetBaseUrl, url), { encoding: opt.encoding ? opt.encoding : "utf8" });
+		return opt.json ? JSON.parse(str) : str;
+	}
+}
+
+export function isHttpProtocol(url: string): boolean {
+	return /^(http|https)\:\/\//.test(url);
+}
