+++ packages/headless-driver/package.json
@@ -10,7 +10,7 @@
     "prepublish": "npm run build",
     "build": "rimraf ./lib && tsc -p ./tsconfig.json && npm run formatter",
     "formatter": "prettier --write \"src/**/*.ts\" --config ../prettier.config.js",
-    "test": "npm run build && npm run run:jest && npm run lint",
+    "test": "npm run run:jest && npm run lint",
     "run:jest": "jest",
     "lint": "tslint -c ../tslint.json -p tsconfig.json -e \"**/node_modules/**\""
   },
@@ -29,14 +29,12 @@
     "@akashic/playlog": "1.3.1",
     "@akashic/trigger": "0.1.5",
     "js-sha256": "0.9.0",
-    "lodash.clonedeep": "4.5.0",
-    "node-fetch": "2.3.0"
+    "lodash.clonedeep": "4.5.0"
   },
   "devDependencies": {
     "@types/get-port": "4.0.0",
     "@types/jest": "24.0.12",
     "@types/lodash.clonedeep": "4.5.4",
-    "@types/node-fetch": "2.1.2",
     "get-port": "^5.0.0",
     "jest": "^24.8.0",
     "prettier": "^1.17.0",
