+++ packages/headless-driver/src/play/Play.ts
@@ -1,9 +1,25 @@
 export type PlayStatus = "preparing" | "running" | "suspending" | "broken";
 
-export interface Play {
+export type Play = (PlayWithContentUrl | PlayWithContentDir) & BasePlay;
+
+export interface BasePlay {
 	playId: string;
 	status: PlayStatus;
-	contentUrl: string;
 	createdAt: number;
 	lastSuspendedAt: number | null;
+}
+
+export interface PlayWithContentUrl {
+	contentUrl: string;
+}
+
+export interface PlayWithContentDir {
+	/**
+	 * game.json を含むディレクトリ。
+	 */
+	contentDir: string;
+	contentConfig?: {
+		externals?: string[];
+		assetBasePath?: string;
+	};
 }
